fix: Break out register fn for shorter syntax
Having to put `{...form.register()}` on a JSX element is clunky, so let
's make it as short and simple as possible by returning the register
function separately from both useCreateForm() and useFormContext().

The register function still exists inside the FormManager of course,
and the old syntax could still be used, but this seems like a simpler
alternative for most use cases.